Split out fetching-related selectors
