+++ client/state/purchases/selectors/fetching.js
@@ -1,0 +1,11 @@
+/**
+ * Internal Dependencies
+ */
+import 'state/purchases/init';
+
+export const isFetchingUserPurchases = ( state ) => state.purchases.isFetchingUserPurchases;
+export const isFetchingSitePurchases = ( state ) => state.purchases.isFetchingSitePurchases;
+export const hasLoadedUserPurchasesFromServer = ( state ) =>
+	state.purchases.hasLoadedUserPurchasesFromServer;
+export const hasLoadedSitePurchasesFromServer = ( state ) =>
+	state.purchases.hasLoadedSitePurchasesFromServer;
